Portability and warning fixes
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@25209 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/proj/get_proj.c
@@ -366,7 +366,7 @@ const char *set_proj_lib(const char *name)
 {
     const char *gisbase = G_gisbase();
     static char *buf = NULL;
-    static int buf_len;
+    static size_t buf_len;
     size_t len = strlen(gisbase) + sizeof(GRIDDIR) + strlen(name) + 1;
 
     if (buf_len < len) {
